training changes

+++ dataloader/clevrer_seg.py
@@ -41,4 +41,4 @@ class CLEVRERSegDataset(Dataset):
         if self.transforms is not None:
             image = self.transforms(image)
 
-        return image, mask
+        return image, mask.long()
